refactor(models): extract Rootable concern
Deduplicate root/parent/children logic shared by Interaction, Task, and Notice.
